Update main.py
The code that generated the first line of the reminder message was using the old formatting when "Reminder" was singular. Fixed the code and made it more elegant such that the same string is used to generate "Reminder" and "Reminders" with the letter "s" added in based on control flow within the f-string. This way, any formatting changes to the first line of the message will be updated for both singular and plural versions.